fix(db): align V2 numeric columns with Double entity mappings
skill_evidence.confidence and skills.years_experience were declared
NUMERIC in the migration but mapped as Double in the entities, which
broke Hibernate ddl-auto:validate on app startup. Use double precision
to match the entity types; the CHECK on confidence is preserved.

+++ backend/src/main/resources/db/migration/V2__evidence_tables.sql
@@ -40,7 +40,7 @@ CREATE TABLE skills (
   normalized_name TEXT NOT NULL,
   category TEXT,
   proficiency TEXT,
-  years_experience NUMERIC,
+  years_experience double precision,
   created_at TIMESTAMPTZ NOT NULL DEFAULT now()
 );
 CREATE INDEX ix_skills_candidate_norm ON skills(candidate_profile_id, normalized_name);
@@ -52,7 +52,7 @@ CREATE TABLE skill_evidence (
     CHECK (source_type IN ('EXPERIENCE','PROJECT','CERTIFICATION','EDUCATION')),
   source_id UUID NOT NULL,
   excerpt TEXT NOT NULL,
-  confidence NUMERIC NOT NULL CHECK (confidence BETWEEN 0 AND 1)
+  confidence double precision NOT NULL CHECK (confidence BETWEEN 0 AND 1)
 );
 CREATE INDEX ix_skill_evidence_skill ON skill_evidence(skill_id);
